Update README.md
Tera Chem is a front-end web application built with HTML, CSS, and JavaScript that allows users to search for different chemicals and their prices. The app features a search bar where users can input chemical names, and it displays relevant information such as chemical properties and pricing. It has a responsive design and interactive elements to enhance user experience.